Activate MariaDB audit plugin and set logging parameters.
Slow query log is activated as well as logging of queries not using indexes. The general log is forced to be disabled to reduce the strain on the server as the MariaDB plugin provides this information. Also, log output is set to file. These changes enable us to use the new built-in functionality in RDS to forward the content of these files directly to CloudWatch.