Performance testing module setup (#26)
Creates a new module for setting up an instance for running performance testing within a specified subnet. Most parameters of the instance are configurable.

Also contains fixes for the issues uncovered during the setup of the new environment.
Namely fixes for overcoming ELB SSL cert catch-22 and bug with elasticache id length.