Disable anacron
We have a cron that runs daily at around midnight.
Amazon Linux comes with anacron which runs at 3am by default.

This is undesirable as it restarts all the docker containers for the
frontend, backend, API and allowed sites.

This creates a race condition as the frontends depend on allowed sites
to pull in configuration