Disable anacron which results in our daily cron scripts being twice Amazon Linux comes with anacron which runs at 3am by default.
This is undesirable as it restarts all the docker containers for the
frontend, backend, API and allowed sites.

This 3am cron also creates a race condition as the frontends depend on allowed sites
to pull in configuration changes.  They are all rebooted at the same
time.

Disable this for Allowed Sites anacron so we just have the original cron
that runs at midnight