Introduce a new variable name for the read replica database
Previously we were overwriting the db hostname with a read replica
value for production.  This was causing trouble for user signups
as we need to insert data.  Inserting data into a read replica
blows up when running that on a read replica.

By separating out the two variables we can specify the read replica
for the authroisations task, and the master for the user signup
task.