Move Manually Created SES Receipt Rules into Terraform
We want this to be managed by Terraform as it involves less manual
effort moving forward.

Creating the rule set itself is still manual as this causes errors with
Terraform applying this accross multiple environments.

Delete the unused test rules which belonged to an inactive test rule set.
Recreate the rules that were manually created.

This doesn't make the new rule set active, it only creates a copy which
can be made active once we are happy.