Create Route53 Latency Alarm
We want to be notified when the Route53 healthcheck latency is more than
1000ms.  This will send a notification to the critical-notifications-arn
endpoint, and we will receive an email about it.

Ultimately we want to receive an sms as well but this will be done in
the console so we don't have to embed our contact details in Terraform.