Change period and evaluation period for scaling actions of ECS Cluster (#92)
Increase time to scale down as this could be triggered during a
deployment.  It is not vital that this happen very soon after we have
additional capacity.

Conversely reduce the delay for scaling up because we want this to
happen as soon as possible when we need additional capacity in the
cluster.