Disable the survey sending behaviour
We're currently not reviewing the surveys received back, instead the
focus of the team is feedback from organisations that are installing
GovWifi into their buildings.

Additionally we have the goal of re-writing the application from PHP
to Ruby.  The surveys code being something that is currently written
in PHP which would need to be rewritten if we want to keep the
functionality.

In the interests of minimising waste, we have decided to not rewrite
the survey behaviour now, and instead re-implement it when we want
to iterate on the end-user experience.  At which point we'd have
capacity inside the team to review, and categorise feedback.