Reduce memory for containers in our API cluster
Currently we have just under 2Gb (1900MB) memory reserved for each container.
The memory utilisation never goes above 5%.
Reduce this memory reservation by half to 950MB per container.

This will save on AWS costs as we need less EC2 instances.
We could reduce this further but would do that
as a separate task after having seen this work in production for a
while.