Pass in the DEVISE_SECRET_KEY to the admin application
The admin application uses Devise, which requires a secret key for
generating hashed tokens for confirmation, password reset and unlock
tokens.  Rotating this key will invalidate all of the above.

We'll reuse the rails secret_key_base which is being used for
encrypting and signing session cookies.

For more information on the DEVISE_SECRET_KEY see:
https://github.com/plataformatec/devise/blob/f220b992c338122226f6fd396056d5a1adf28df8/lib/generators/templates/devise.rb#L11