Populate Allowed Sites API environment variables
We need access to the admin database in the allowed sites API.
This is so we can populate the radius configuration from 2 datasources
as we are in the process of going live with the admin.

Also delete the db-instance-count variable which was unused and defaults
to 1.  This was raising a warning.