Create Device Wifi Frontend Infrastructure for Staging
This creates a new Radius frontend which has its own ECR docker
repository.

It is currently a clone of our other frontends, but I imagine that this
setup would deviate a lot as we move forward.

It has an Elastic IP associated with this frontend, which we can use to
configure our SSID to point to this to do the certificate based
authentication.