Rolling restarts for docker containers
We install security updates each night.
The time at which this happens is determined by the last numbers of the ip % 59
This is fragile and is triggering health check alarms.

Instead of the heavy handed `sudo service docker restart`, do a rolling
restart, which is faster and should ensure that there is always a task
running to serve traffic.