Fix performance platform logging statistics
We switched to Fargate for the Logging API.
Turns out that scheduled tasks with Fargate are not supported in London yet.

https://docs.aws.amazon.com/AmazonECS/latest/developerguide/scheduled_tasks.html

This means that we have to continue to run our scheduled tasks on EC2
instead of Fargate.  This introduces some complexity as we need 2 task
definitions to achieve this.

One for ECS Tasks and one for ECS Scheduled tasks.

This is temporary until Fargate for Scheduled events is available in
London.