Extract new user details database
We are pulling the end user details table out of the main database.
This database currently has sessions and user details in it.
User details data is critical while session data is just used for
statistics.

This also introduces a new "env" variable which will replace "rack-env"
so it can be used in more contexts.