Remove burst balance alarm for user database (#202)
We initially set this alarm up for the sessions database which was
having trouble with the read replica lagging behind by about 10 minutes.

We found that this lag was tied to the available burst balance credits.

We don't need to monitor this for the users database as it doesn't
contain as much data.  Replicating will be fast.