Remove deprecated tasks (#247)
We just removed the `publish_daily_statistics` rake task from the
logging api as we don't need it any more. However, the ECS scheduled
task is still running, and can't find the rake task. We need to delete
this scheduled task from the infrastructure.

This is causing throwing a RuntimeError caught by Sentry and false
alarming our Slack channel.