Adjust radius to auth api alerting threshold
From: 1 error message detected in the last 5 minutes
To: 1 error message detected in each of the last two minutes

An occassional error is fine (it seems to occur ~9:30am and just after lunch which suggests it is a load based), and it quickly heals upon a subsequent attempt to authenticate. We are more interested when there is a sustained period of errors indicating a real problem which is what this change to threshold allows.