REP-461: Use all AZs in London
eu-west-2 now has three availability zones, so let's use them. The
changes in this commit need to be applied in two stages. First, apply
the backend module changes:

    make terraform_flags="-target=module.backend" <env> init-backend apply

Then you can go ahead with:

    make <env> apply

You might have to run this a couple of times so that the Route53
healthchecks get sorted out.