REP-462: Use separate container to populate cert config
Add a one-off `raddb` container to the frontend task definition and make the
frontend container depend on successful completion of the `raddb`
command. The two containers share a volume which is mounted to
`/etc/raddb/certs`. Since the client config is pulled from S3 as well,
change the `$INCLUDE` to pull `clients.conf` from `certdir`.