Fix dependency ordering problem preventing S3 bucket creation.
The api and frontend modules depend on the admin S3 bucket, but there
was nothing in the code to ensure that the bucket gets created before it
is used.

Add an output to the admin module (which creates the bucket) and use
that output instead of hardcoding the bucket name in main.tf. This way,
Terraform knows to create the bucket before using it.