Increase RDS Storage alert from 20GB to 30GB
The alerts are firing because storage has passed 20GB.
This is fine and there are no concerns about its growth
to this point.

The RDS instances have 1TB storage assigned, and these
alerts are just used to make us aware if there is
sudden and unplanned growth in storage.

Solo: @smford