fix: add an IAM user to access the dashboard bucket
Because our dashboard app is not terraformed with the rest of our
infrastructure (it lives on PaaS), use an intermediate IAM role whose
credentials we will generate through the AWS console, then store in
govwifi-build, then encrypt within the dashboard app.

This is directly inspired by the `govwifi-api/wordlist.tf` file and
its jenkins user which can be consulted for further reference.