Remove useless rds-kms-key-id variable
We have removed the rds-kms-key-id variable as it is not needed because AWS manages it entirely unless we specifically create one ourselves - which we don't do.
This should work for new-staging, but we need to be cautious on how this change effects production.  To test how it would behave on production, we should run a plan and apply with the 'rds-kms-key-id' removed against old-staging.