Junk default values for staging-london-temp certificates
These are junk values to help us surface issues in the terraform output and aid is in solving the problem of
how to manage the acm certificates and their binding to their matching
elb.

The variables elb-ssl-cert-arn and govwifi-api-ssl-cert-arn were
previously defined explicitly within secrets.auto.tfvars (from govwifi-build), which causes a chicken and egg situation to arise.  Instead, we should be managing the elb-acm relationship automatically rather than doing a manual association.  However, a change to this would result in a change to production.  This is tech debt we should address in the future, but until that is solved properly we will need to transplate the certificates in by hand.