Bodge:Define name of the route53zone for staging-london-temp
Tech Debt alert.

This is a bodge that should not be merged on to master, that allows us
to provision a new environment (staging-london-temp) from a delegated
route53 zone.

The route53 zone for staging-temp.wifi.service.gov.uk exists within the
new staging aws account.  For it work we have had to manually create the
zone and then delegate from route53 hosted in production to this newly created zone.

Since old-staging and production sit within the same aws account and can
access the same route53 zone, they happily work with `name = "wifi.service.gov.uk."`.

However, this will not work in the new world where there is an
environment per aws account.  It is related to the problem of the naming
of environments and is complex in nature.  The problem being the naming
patterns between production environment (www.wifi.service.gov.uk) and all
other environments (www.staging.wifi.service.gov.uk) is inconsistent,
combined with terraforms lack of conditional statements, makes it hard
to manage.

The proper solution to this problem is simply not having the route53 zone baked into a terraform
module common to all environments, and instead pass it in after being
interpolated at a high level.  The problem with fixing it in this manner is that it will effect production and is
not testable as there is no environment setup the same way as
production.

This problem needs to be somehow addressed before merging in to master,
this commit implements a bodge to allow us to work on other tasks.