Set Storage Time For Prometheus Data To 90 Days
We want to keep the Prometheus log data for a reasonable period
of time to facilitate debugging and recognise patterns. However
we don't want to keep it forever. We can easily adjust the length
of time we keep the data for in the future.

I have followed the guidance in the official documentation
on how this value should be set: https://prometheus.io/docs/prometheus/latest/storage/

Interestingly it is NOT set in the Prometheus config file. See here for more information:
https://stackoverflow.com/questions/59298811/increasing-prometheus-storage-retention