Add Customer Managed KMS Key For Encrypting Database Snapshots
**WHAT**
Add a customer managed KMS key for encrypting database snapshots. At present this
is only used to encrypt the primary staging account snapshots and data at rest. The secondary
staging account has access to this key and will use it to sync it’s data with the primary
staging account.

**WHY**
In order to perform the data sync between accounts, we need to share access to the keys which
requires a change in the key policy. The current RDS KMS key is managed by AWS, it is not possible
to edit policies on AWS managed keys (and therefore not possible to grant access to other accounts),
so a new KMS customer managed key needed to be created and used to encrypt the Staging database snapshots.