Temporarily Disable Backup Script
Comment out backup script cron job in user data for now. Another backup script has been created here:
https://github.com/alphagov/govwifi-terraform/pull/436/files

This uses an ECS task instead of a shell script. It is probably better for us to use the ECS task in the long run, and switching to it will mean that Secondary Staging stays more in sync with production.