Setup SNS things for PagerDuty in all three production regions
This is so Cloudwatch alarms in these regions can send events through
to PagerDuty, via the SNS topic and subscription.

The HTTPS endpoint configuration for the subscription is handled as a
secret. This exists in eu-west-2, and is replicated to eu-west-1 to
allow Terraform to look it up in that region too.