Add commented out Terraform for composite alarms for Radius servers
Out of hours, someone needs to be notified if all three Radius servers
in a region are down, which is the same as saying that the healthcheck
alarms for each Radius server in a region is not OK. This is what this
alarm checks.

A much newer version of the AWS provider is needed to enable this
Terraform, so it's being added now but commented out. The alarms in
eu-west-1 and eu-west-2 have been created manually.