Add alarms for high level coverage of some important services
This is part of the work to enable out of hours support for
GovWifi. There are alarms already for one unhealthy host connected to
a load balancer, but out of hours, it's important to just alarm if
there's no healthy hosts connected to a load balancer.

That's what these alarms do. They're connected up to the SNS topic for
notifying PagerDuty.