Update the AMIs used in govwifi-frontend
Switch from using the Amazon Linux AMI to Amazon Linux 2, this changes
a few things, like switching to systemd as the init system for
example.

As part of this, update and try to simplify the user data. Remove the
CloudWatchMonitoringScripts as they're deprecated, and I'm not sure
what they do. Switch to using yum-cron to provide security
updates. Stop breaking anacron, as I think this change is probably
unnecessary with the switch to yum-cron (as it had something to do
with cron.daily [1]). Remove the upstart service for starting the
awslogs service, just set the configuration correctly and then start
it with the script.

1: https://github.com/alphagov/govwifi-terraform/commit/34f7f1e012c7b396aa591adf8abc4c4d11953125