Tighten egress from the grafana-alb-out security group
var.bastion-ips is actually the singular IP address of the bastion EC2
instance, plus the /24's for each of the availability zones of some
VPC. This isn't great, since it's necessary as the Grafana EC2
instance is covered by the rules, but they're overly permissive.

This change tightens up the security rules to only allow egress to the
security group for the Grafana EC2 instance. It has to be defined as a
separate resource to avoid creating a cycle between the two security
groups involved here.