Address wierdness with the bastion-ips variable for govwifi-grafana
The description says it's a list of hosts, but it's actually a list of
CIDR blocks, one of which is a /32 containing the IP address of the
bastion machine, but there's some other larger blocks as well.

All that's required for the grafana module is just the IP address of
the bastion machine, so move things in that direction.

Unfortunately, var.bastion-server-IP is also not actually an IP
address, but a single CIDR block :( so fix that when passing the value
to the module.