Refactor a couple of variables in the govwifi-backend module
Previously, the govwifi-backend module took the following variables:
bastion-server-IP, bastion-server-ip, and backend-subnet-IPs. These
changes refactor it to instead take these two variables:
bastion_server_ip and backend_subnet_cidr_blocks.

This addresses the duplication of bastion-server-ip and
bastion-server-IP, and also that these variables didn't contain IP
addresses, but instead a CIDR block.

Similarly, backend-subnet-IPs was a comma separated list of CIDR
blocks. It's been changed to backend_subnet_cidr_blocks, which is a
list of those same CIDR blocks.

The variable names and values are still confusing where the module is
used, but at least within the module things should be a bit
better. Hopefully in the future, the naming and values within
govwifi-build can be improved similarly.