Sort out var.frontend-radius-IPs
This variable comes from govwifi-build. It's not a list of IPs, but
instead a comma separated string of CIDR blocks.

To improve on this situation, create a new local called
frontend_radius_ips, which is actually a list of IPs from the
combination of var.london-radius-ip-addresses and
var.dublin-radius-ip-addresses.

Where modules where using var.frontend-radius-IPs, they'll now be
using local.frontend_radius_ips in one form or another. Some changes
within modules have been made so they handle this being a list of IP
addresses.

This change shouldn't result in any actual changes, it's just
refactoring the Terraform.