Tweak the frontend fe-admin-in security group
In most environments, no change expected in staging and
staging-dublin-temp.

Rather than passing a list of bastion-ips to govwifi-frontend, just
pass the singular bastion IP. I'm not sure why, but in most
environments the backend-subnet-IPs where also on this list. I believe
that's a capability that can be removed. I don't think everything in
those IP ranges needs acesss.