Change variables coming from govwifi-build
This commit changes the names and many expected values of all of the
variables coming from govwifi-build, for all 6 current
environments. It's intended to be merged alongside a corresponding
change in govwifi-build which changes the names and many variable
values.

There are two things I'm trying to improve with this commit. Variable
names (underscores rather than hyphens), and having the variable name
accurately describe the values (so bastion_server_ip should be an IP
address for example).

To avoid needing to coordinate many merges between govwifi-build and
govwifi-terraform, I've grouped this all together in one large
commit. Hopefully such a change won't be necessary again.