Increase the govwifi-api low_cpu_threshold by 10x
This affects both the logging API and authentication API. I'm actually
just looking at the logging API, but the previous value of 1 looks to
be far too low. The CPU usage will never get below it, so the service
will never scale down.

Raising this to 10 should mean both services actually scale down when
they're not under much load (although CPU utilisation is not a great
metric to scale by in general).