Refactor resources related to the GovWifi backups
Refactor resources related to the GovWifi backups so there is less hardcoding.
Replace Env-Name with Subdomain-Name in the S3 bucket name (s3 buckets must be
unique, Env-Name is not always a unique variable but the subdomain will be).

Also Enable RDS to S3 backups in new staging environment. As we are now passing
the rds_mysql_backup_bucket as a variable instead of hardcoding it, we also
need to make changes to the main.tf files in wifi-london and wifi.