Upgrade the Terraform AWS provider to version 3.63.0
From version 2.10.0. Keeping the provider up to date is useful, and
specifically this should allow managing the composite alarms with
Terraform.

This commit removes the version constraint in the main.tf files, the
lockfile stores the exact version and I don't want a specific one, so
just let Terraform sort it out. I ran terraform init -upgrade to
update the lockfiles.

The only other necessary change from this upgrade is changing some of
the aws_alb_listener_rule resources.