Change the action for the radius_cannot_connect_to_api alarm
To use the route53-critical-notifications topic, as this seems more
appropriate than the devops one (although I'm not sure what either are
for).

The actual motivation for this is to hear about this alarm going off
in Slack, and the route53-critical-notifications topic is hooked up to
Slack already, so this change should mean that we're notified in Slack
when this alarm goes off.