Rename the govwifi-admin data S3 bucket name output
I don't think there's a need to say it's on AWS, or that it's "admin"
related, since this is the admin module. Instead specify that it's
related to the app, and that it's the name of the S3 bucket.

This output is currently unused as far as I'm aware, it was used
briefly in Terraform, but then replaced by a hardcoded value. I'm
changing this as part of going back to feeding the value through
Terraform to where it should be used.