Refactor how the admin app data S3 bucket name is handled
Rather than hardcoding the name of the bucket in the Terraform for
each environment, pass the value around instead. This should make the
Terraform easier to read.

This is a little more complicated because the bucket name is used in
both the London and Dublin environments, but the resource lives in the
London environment (as the admin app only runs in London). To pass the
value between environments, make it an output of the environment
itself in the London environments and then access the value in the
Terraform state for the London environments from the respective Dublin
environment.

In addition to the above changes, the naming has changed as well, as I
think names like admin_app_data_s3_bucket_name is more informative
than admin-bucket-name.