Fix a mismatch between AWS regions and SNS topic notification ARNs
In the govwifi-frontend module. This commit
d09a5913017955d3faf7535c94cac9f01342dab9 changed the
radius_cannot_connect_to_api alarm action, but this caused a problem
as the topic used was in us-east-1, which can't be hooked up to the
alarm in the eu-west-1/eu-west-2 regions.

Now that I've seen this, I think I can see that the
route53-critical-notifications-arn should probably be called
us_east_1_critical_notifications_arn. It was probably called route53
because route53 is managed from us-east-1, but actually saying
us-east-1 makes so much more sense.

This commit makes that change, as well as adding a
critical_notifications_arn variable, which should be passed the
appropriate SNS topic ARN for the current region.