Don't specify the subnet_id for the Grafana instance
Count will either be 0 or 1, in the 0 case, there isn't any instance
to create, so this looks to me like a complicated way of sticking the
Grafana instance always in the first subnet.

I don't think it needs to be in a specific subnet, so just don't
specify this optional argument.

I'm looking at getting rid of the count entirely, which requires doing
something about this.