Tweak the Logging API resources
Increase the CPU to 512 (1/2 a vCPU) and increase the memory
accordingly to the lowest possible value (1024).

The intent here is to increase the CPU resource available, in the hope
that this makes the service more stable and scalable.

I'm not quite sure why each of the values are set out twice in the
Terraform.