Remove references to "temp" in staging step one
Update staging main.tf and variables.tf files to remove references to temp
(this was a useful distinguisher when we had two staging environments, it is
now no longer necessary). Leave the names of the KMS/ssh keys, S3 terraform
state bucket and directories as they are for now (they will be changed in a
later commit to avoid breaking things).