Move the US East 1 pagerduty endpoint to be managed by wifi-london
Rather than wifi.

This is in preparation for managing the composite alarms through
Terraform. It will involve passing the topic-arn output by this module
from one environment to the other, and since there's already some
remote state access from wifi to wifi-london, I think it's better to
keep the shared things in wifi-london (so that the remote state
accessing is just going one direction).