Manage the composite alarms with Terraform
The composite alarm resource was added in
5cd79c9dbfc18b2c60645064927f47d644b49e74, but commented out and
instead was created manually in AWS.

It wasn't feasible to manage through Terraform at the time because the
version of the AWS provider was insufficient, and there was no pattern
for sharing the topic ARN across environments.

Both those issues have now been addressed, so these alarms can now be
managed via Terraform.