Convert some variables to locals in the wifi environment
For three variables that don't (and shouldn't) vary.

Variables form part of the interface for Terraform configuration, and
as is in the name, they're usually used for things that vary. The
env_name, env_subdomain and product_name for the wifi environment
Terraform configuration aren't things that vary, so I think it's
clearer to make them locals, rather than variables.