Remove decryption code from the unencrypt-secrets script
As it's no longer used. I think this has been the case for a while,
since the move to storing secrets in AWS.

This script could be renamed to copy-secrets, but I think it might be
better to try and think if the relationship between govwifi-terraform
and govwifi-build can be broken (moving the data their either in to
govwifi-terraform or AWS).