Set health_check_grace_period_seconds for API ECS services
I was looking at this when there were issues in Production for the
Logging and Authorisation APIs, tasks kept coming and going, and I
think the lack of a grace period for the health check may have led to
this behaviour.

The value of 20 seconds is just an initial guess.