Tighten up the health checks for the API service ALBs
Having a ALB instance wait for 10 consecutive failures at intervals of
10 seconds is far too long, that would mean that a target would still
be used for up to 100 seconds after it started timing out.

I think 2 consecutive failures is a much better value, as that should
result in unhealthy targets being removed faster. Decrease the timeout
as well, 4 seconds is probably still too high, but I just wanted to
decrease it a bit.