Simplify the handling of the Route53 zones
Currently the zone is created by hand, and not managed through
Terraform. The zone ID is discovered in two different ways, either
through a data source, or a secret.

This is quite confusing as it makes it hard to see where the zone is
used, and makes it difficult to use a Terraform resource to manage it.

I think managing the zone with a resource is a good end goal, this
commit just uses one single data source to lookup the zone ID, and
then passes this around to where it's needed. The secrets can then be
deleted.