Change how subnets are handling in govwifi-backend
Rather than passing in the zone_count, zone_names and zone_subnets,
lookup the availability zone details from AWS, and derive the subnet
blocks for the VPC CIDR block.

I think there are a few advantages to this. There are less variables
to manage, which simplifies the configuration, instead values looked
up or derived from the appropriate place. This is important in
Terraform because it informs resource dependencies.

Applying this without re-creating the subnets will require moving the
old subnets to the new places in the statefiles.