Replace wifi_frontend_subnet with frontend_subnet_id
These two outputs are identical. I'm looking at refactoring the
subnets, and rather than having to change how both of these are
constructed, I'd rather just change one.

I believe wifi_frontend_subnet was added second in [1], so change that
to frontend_subnet_id. The name of the output could be improved, but
that's a different task.

1: 35746b77e42fe4d491abeaaafb6c6020e221a4e6