Address IAM duplication between govwifi-account and govwifi-api
Things like the read_wordlist_policy look to have been copied in to
govwifi-api, but also left in govwifi-account, with the
is_production_aws_account variable also added to avoid Terraform
trying to create the same policy twice in production.

Remove the use of is_production_aws_account by removing IAM things
that look to be duplicated in govwifi-account. This is a little
confusing since govwifi-account looks to grant access via a group,
whereas govwifi-api manages it directly with users.