Set is_production_aws_account to default to true in govwifi-api
In this first pass of trying to get rid of is_production_aws_account,
I'm going to avoid dealing with the wordlist bucket, since it has a
name in production which doesn't include the environment. Having a
default means at least some of the use of the variable can be removed.