Re-add some policies to the deploy smoketest user
These were removed from the GovWifi-Pipeline group in [1], which the
govwifi-pipeline-deploy-smoketest user is a member, which broke the
push-to-ecr bit of the smoke-tests pipeline in Concourse.

1: 17d908d852ab353bfae34b14a65364be747ca214

I've added these back, but attached specificly to the user, rather
than the group. This makes it much clearer where they are needed.

Hopefully in the long term, this stuff related to smoke-tests can be
extracted in to an isolated bit of Terraform, maybe a module.