Switch the wordlist bucket to use a bucket prefix
This is a rigorous way of ensuring the buckets have different names in
different environments, without having to put the environment or
subdomain in to the bucket name.

This bucket is used in the deploy pipeline, so after this change is
applied, the pipeline will need updating with the name of the
bucket. This name will be a little ugly compared with the previous
name, but I think this is OK, the name shouldn't change regularly.