Fix the frontend cert bucket access for govwifi-deploy-pipeline
Rather than assuming this is not production and building the bucket
ARN with the env_subdomain, move the policy to the govwifi-frontend
module so that the actual ARN for the bucket can be easily accessed.

This probably should be in neither module, and something associated
with the pipelines, but that doesn't exist yet.