Move Staging GovWifi Email Module To London (eu-west-2 region)
This same change will be rolled out in production shortly.

The reasons for this change are:

- Most of GovWifi's critical infrastructure is in the London region already. It
  is jarring that SES and SNS resources are in Ireland. When the original
designers of GovWifi set up the infrastructure, they started building it in
Ireland but when Brexit occurred they switched the critical components to the
London region (the plan was to only use resources in Ireland as a backup if for
some reason the London region went down). At the moment the SES and SNS
resources only exist in Ireland (they are acting as primary resources not
backups). This is not inline with the pattern of the rest of our
infrastructure.

- Currently our AWS, SES and SNS resources are in the Ireland region of AWS. In
  order to complete this https://trello.com/c/QSO59546/1889-lockdown-requests-coming-into-the-user-signup
we need them to be in the London region.