Delete some duplicated roles from the govwifi-account module
Duplicated Terraform resources are bad, and I believe in this case
Terraform was tricked in to managing the same roles in AWS twice.

For deploying this, my plan is just to remove the Terraform state. If
I'm correct and these are all duplicated resources, there will be no
changes.

I've split this across multiple commits as that seemed to make the
diff cleaner.