Tweak how the Grafana EIP is managed in Terraform
Remove the instance argument from the aws_eip, as the association with
the instance is done through the aws_eip_association, and this change
is important to avoid cyclic dependencies between Terraform resources.

Also remove the redundant depends_on bit from the aws_eip_association,
and expose the public IP through an output.