Have Terraform manage the EIP used for the bastion
Rather than creating it manually. This means that Terraform can look
up the public_ip associated with the EIP, rather than having it passed
in as a variable.

Unfortunately the bastion_server_ip variable can't be totally removed
yet, since in staging-dublin, the bastion IP actually comes from the
other region. This can be addressed once the Terraform for staging is
consolidated.