Add a new-terraform-state module
This is a module for handling an S3 bucket to store Terraform
state. It's an adaptation of the existing terraform-state module,
changed to take two AWS providers, representing the primary region and
the region to use for replication. Thus making it suitable for use
when the Terraform for a single govwifi deployment is consolidated.