Consolidate the Terrraform for staging
Rather than having the Terraform split in to two bits, roughly
corresponding to the two main regions used, have one consolidated set
of Terraform that works across all relevant regions.

This has a number of advantages, both for general use of Terraform,
but also for bootstrapping new deployments of GovWifi.

Having the Terraform split in to two lumps meant it was hard to share
values and resources, which has led to things like overusing EIPs,
duplicating IAM resources, and managing some resources twice, once
from each "environment".

In the case of bootstrapping, separating the Terraform roughly by
region meant that coordination of creating some things has to be done
manually, since Terraform could not reason about some resource
dependencies. Consolidating the Terraform addresses this.

To achieve the consolidation with minimal effort, a script has been
used to consolidate the Terraform state files. This means there's very
few changes that Terraform wants to make. The main change I see is
that the unused EIPs for the bastion and prometheus machines in the
dublin region aren't being included in security groups any longer.