Mark the docker_image_path values as nonsensitive
This secret actually contains the URL for a AWS ECR
repository. Ideally, this wouldn't be a secret at all, and the
Terraform would just describe where the attribute from the repository
resource should be used.

Unfortunately, this is difficult to address until the Terraform for
different regions is consolidated, so for now, just mark the values as
non sensitive to improve the plan output.