Ensure the daily metrics logging task runs after the set_user_last_login task
The code that collects daily data on user signups that have also logged in
depends on the last_login field being set. It is therefore necessary
that this code should run after the daily_gdpr_set_user_last_login has run
which sets this field daily from the sessions database.