Start the transition to frontend using fargate and a load balancer
The commit adds a whole new set of resources to run frontend (RADIUS)
as a load balanced service using ECS Fargate.

The intent here is to run this in addition to the current EC2 setup,
then migrate EIPs over to the load balancer. For testing prior to
this, a new set of EIPs is used, and these will be deregistered once
the migration is complete.