Rename all container references 'logging' to 'logging-api'
Make container name consistent with the name we use for the ECS service,
cluster and task definition. In our ECS task definitions we currently declare a
container name that is different from the app, task and cluster name:
https://github.com/alphagov/govwifi-terraform/blob/13c45e4791ba82dfaba89cf68847eb2d720c4825/govwifi-api/logging-api-cluster.tf#L144

This name is also referenced in our application code, see here for an example:
https://github.com/alphagov/govwifi-logging-api/search?q=container_name

The name used is "logging" but elsewhere we use "logging-api" to refer
to related resources like the ECS service. This should be consistent -
"logging-api" should be used everywhere, and references to "logging"
should be deleted.

Why
Using different values for the container name, service name and cluster
name is confusing and makes writing AWS pipeline code more complicated.